+++ src/Services/ResourceService.php
@@ -188,7 +188,7 @@ class ResourceService
     /**
      * @param AbstractHeidelpayResource $resource
      *
-     * @return null
+     * @return AbstractHeidelpayResource|null
      *
      * @throws HeidelpayApiException
      * @throws \RuntimeException
@@ -761,6 +761,67 @@ class ResourceService
         $webhook->setParentResource($this->heidelpay);
         $this->create($webhook);
         return $webhook;
+    }
+
+    /**
+     * Updates the given local Webhook object using the API.
+     * Retrieves a Webhook resource, if the webhook parameter is the webhook id.
+     *
+     * @param Webhook|string $webhook
+     *
+     * @return Webhook
+     *
+     * @throws HeidelpayApiException
+     * @throws \RuntimeException
+     */
+    public function fetchWebhook($webhook): Webhook
+    {
+        $webhookObject = $webhook;
+        if (\is_string($webhook)) {
+            $webhookObject = new Webhook();
+            $webhookObject->setId($webhook);
+        }
+
+        $webhookObject->setParentResource($this->heidelpay);
+        $this->fetch($webhookObject);
+        return $webhookObject;
+    }
+
+    /**
+     * Updates the Webhook resource of the api with the given object.
+     *
+     * @param Webhook $webhook
+     *
+     * @return Webhook
+     *
+     * @throws HeidelpayApiException
+     * @throws \RuntimeException
+     */
+    public function updateWebhook($webhook): Webhook
+    {
+        $webhook->setParentResource($this->heidelpay);
+        $this->update($webhook);
+        return $webhook;
+    }
+
+    /**
+     * Deletes the given Webhook resource.
+     *
+     * @param Webhook|string $webhook
+     *
+     * @return Webhook|AbstractHeidelpayResource|null
+     * @throws HeidelpayApiException
+     * @throws \RuntimeException
+     */
+    public function deleteWebhook($webhook)
+    {
+        $webhookObject = $webhook;
+
+        if (\is_string($webhook)) {
+            $webhookObject = $this->fetchWebhook($webhook);
+        }
+
+        return $this->delete($webhookObject);
     }
 
     //</editor-fold>