+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 *   Added PublicKeyValidator.
+*   Added Webhook resource.
 
 ### Fixed
 *   Several code style issues.